Attempt to make cmake work (didn't work)

+++ src/mGL/MeshFactory.cpp
@@ -8,37 +8,37 @@
 #include "Renderable/RenderableModel.h"
 #include "Mesh.h"
 #include "MeshFactory.h"
-#include "mBase/StringUtils.h"
+#include "StringUtils.h"
 #include "Vertex.h"
 #include "Material/MaterialFactory.h"
-#include "mBase/Logger.h"
+#include "Logger.h"
 
-#include "mBase/XML/tinyxml2.h"
+#include "XML/tinyxml2.h"
 
 namespace mGL
 {
-	RenderableObject* MeshFactory::LoadObj(const std::string& path)
+	RenderableObject *MeshFactory::LoadObj(const std::string &path)
 	{
 		auto meshes = GetMeshes(path);
-		RenderableObject* ro = new RenderableModel(meshes);
+		RenderableObject *ro = new RenderableModel(meshes);
 		return ro;
 	}
 
-	RenderableObject* MeshFactory::LoadObj(tinyxml2::XMLElement* element)
+	RenderableObject *MeshFactory::LoadObj(tinyxml2::XMLElement *element)
 	{
-		const char* cpath;
-		if (element->QueryStringAttribute("model", &cpath) != tinyxml2::XML_SUCCESS) return nullptr;
+		const char *cpath;
+		if (element->QueryStringAttribute("model", &cpath) != tinyxml2::XML_SUCCESS)
+			return nullptr;
 		std::string path = std::string(cpath);
 		auto meshes = GetMeshes(path);
-		RenderableObject* ro = new RenderableModel(element, meshes);
+		RenderableObject *ro = new RenderableModel(element, meshes);
 		return ro;
 	}
 
-	std::vector<Mesh*> MeshFactory::GetMeshes(const std::string& path)
+	std::vector<Mesh *> MeshFactory::GetMeshes(const std::string &path)
 	{
 		if (path.substr(path.size() - 4, 4) != ".obj")
 			return {};
-
 
 		std::string fullPath = "Data/Meshes/" + path;
 		Logger::Log("Mesh Factory", "Trying to open obj file on path " + fullPath);
@@ -59,11 +59,12 @@ namespace mGL
 		std::unordered_map<Vertex, unsigned short> indicesMap = std::unordered_map<Vertex, unsigned short>();
 		std::unordered_map<std::string, std::shared_ptr<Material>> materialMap;
 		std::shared_ptr<Material> currentMaterial = nullptr;
-		std::vector<Mesh*> meshes = std::vector<Mesh*>();
+		std::vector<Mesh *> meshes = std::vector<Mesh *>();
 		unsigned short indexCount = 0;
 		while (std::getline(file, line))
 		{
-			if (line.substr(0, 1) == "#") continue;
+			if (line.substr(0, 1) == "#")
+				continue;
 			std::string firstToken = mBase::Strings::FirstToken(line);
 			if (firstToken == "mtllib")
 			{
@@ -92,8 +93,9 @@ namespace mGL
 			}
 			else if (firstToken == "usemtl")
 			{
-				if (currentMaterial != nullptr) {
-					Mesh* mesh = new Mesh(vertexs, indices);
+				if (currentMaterial != nullptr)
+				{
+					Mesh *mesh = new Mesh(vertexs, indices);
 					mesh->SetMaterial(currentMaterial);
 					meshes.push_back(mesh);
 				}
@@ -115,7 +117,7 @@ namespace mGL
 				std::vector<std::string> faces;
 				mBase::Strings::Split(mBase::Strings::Tail(line), faces, " ");
 				unsigned int firstIndex;
-				for (int i = 0; i < faces.size(); ++i)
+				for (size_t i = 0; i < faces.size(); ++i)
 				{
 					std::vector<std::string> splitted;
 					mBase::Strings::Split(faces[i], splitted, "/");
@@ -125,7 +127,8 @@ namespace mGL
 					Vertex v = Vertex();
 					v.Position = positions[positionIndex - 1];
 					v.Uvs = uvs[uvIndex - 1];
-					if (splitted.size() > 2) {
+					if (splitted.size() > 2)
+					{
 						unsigned short normalIndex = std::stoi(splitted[2]);
 						v.Normal = normals[normalIndex - 1];
 					}
@@ -140,7 +143,8 @@ namespace mGL
 						if (faces.size() == 4 && i == 3)
 							indices.push_back(firstIndex);
 					}
-					else {
+					else
+					{
 						// Not exists. Create new one
 						vertexs.push_back(v);
 						indicesMap[v] = indexCount;
@@ -156,8 +160,9 @@ namespace mGL
 				}
 			}
 		}
-		if (currentMaterial != nullptr) { // There is no indicator of EOF, so on the end we use all data for the last material configured
-			Mesh* mesh = new Mesh(vertexs, indices);
+		if (currentMaterial != nullptr)
+		{ // There is no indicator of EOF, so on the end we use all data for the last material configured
+			Mesh *mesh = new Mesh(vertexs, indices);
 			mesh->SetMaterial(currentMaterial);
 			meshes.push_back(mesh);
 		}
